+++ sdk/search/arm-search/src/api/adminKeys/options.ts
@@ -1,0 +1,16 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+import type { OperationOptions } from "@azure-rest/core-client";
+
+/** Optional parameters. */
+export interface AdminKeysRegenerateOptionalParams extends OperationOptions {
+  /** A client-generated GUID value that identifies this request. If specified, this will be included in response information as a way to track the request. */
+  clientRequestId?: string;
+}
+
+/** Optional parameters. */
+export interface AdminKeysGetOptionalParams extends OperationOptions {
+  /** A client-generated GUID value that identifies this request. If specified, this will be included in response information as a way to track the request. */
+  clientRequestId?: string;
+}